pyproject.toml
[project]
name = "openags"
version = "0.2.0"
description = "Open Autonomous Generalist Scientist — autonomous research agent framework"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["agent", "research", "llm", "autonomous", "science"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.9",
    "rich>=13",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "litellm>=1.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "httpx>=0.24",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
    "mypy>=1.0",
]
pdf = [
    "pymupdf>=1.23",
]
rag = [
    "chromadb>=0.4",
]

[project.scripts]
openags = "openags.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
